Add script to refresh lineage test snapshots.
