+++ src/antlr/lineage.tests/ORDER BY references.json
@@ -3,165 +3,169 @@
   "options": {
     "positionalRefsEnabled": true
   },
-  "data": [
-    {
-      "type": "table",
-      "id": "result_2",
-      "label": "account",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 46,
-        "endColumn": 53
-      },
-      "data": {
-        "id": "account"
-      },
-      "isSourceOnly": true,
-      "columns": [
-        {
-          "id": "account_type",
-          "label": "account_type",
-          "data": {
+  "data": {
+    "nodes": [
+      {
+        "type": "table",
+        "id": "result_2",
+        "label": "account",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 46,
+          "endColumn": 53
+        },
+        "data": {
+          "id": "account"
+        },
+        "isSourceOnly": true,
+        "columns": [
+          {
             "id": "account_type",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_id",
-          "label": "account_id",
-          "data": {
+            "label": "account_type",
+            "data": {
+              "id": "account_type",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_id",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_description",
-          "label": "account_description",
-          "data": {
+            "label": "account_id",
+            "data": {
+              "id": "account_id",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_description",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_parent",
-          "label": "account_parent",
-          "data": {
+            "label": "account_description",
+            "data": {
+              "id": "account_description",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_parent",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_rollup",
-          "label": "account_rollup",
-          "data": {
+            "label": "account_parent",
+            "data": {
+              "id": "account_parent",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_rollup",
-            "tableId": "account"
+            "label": "account_rollup",
+            "data": {
+              "id": "account_rollup",
+              "tableId": "account"
+            }
           }
-        }
-      ]
-    },
-    {
-      "type": "table",
-      "id": "result_1",
-      "label": "[final result]",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 0,
-        "endColumn": 86
+        ]
       },
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "aid",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 7,
-            "endColumn": 21
-          }
+      {
+        "type": "table",
+        "id": "result_1",
+        "label": "[final result]",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 0,
+          "endColumn": 86
         },
-        {
-          "id": "column_2",
-          "label": "ap",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 23,
-            "endColumn": 40
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "aid",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 7,
+              "endColumn": 21
+            }
+          },
+          {
+            "id": "column_2",
+            "label": "ap",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 23,
+              "endColumn": 40
+            }
           }
+        ],
+        "isTargetOnly": true
+      }
+    ],
+    "edges": [
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_id"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_1"
         }
-      ],
-      "isTargetOnly": true
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_id"
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_1"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_parent"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_parent"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_2"
+        }
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_2"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "group by",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_parent"
+      {
+        "type": "edge",
+        "edgeType": "group by",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_parent"
+        },
+        "target": {
+          "tableId": "result_1"
+        }
       },
-      "target": {
-        "tableId": "result_1"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "group by",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_id"
+      {
+        "type": "edge",
+        "edgeType": "group by",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_id"
+        },
+        "target": {
+          "tableId": "result_1"
+        }
       },
-      "target": {
-        "tableId": "result_1"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "order by",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_parent"
+      {
+        "type": "edge",
+        "edgeType": "order by",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_parent"
+        },
+        "target": {
+          "tableId": "result_1"
+        }
       },
-      "target": {
-        "tableId": "result_1"
+      {
+        "type": "edge",
+        "edgeType": "order by",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_id"
+        },
+        "target": {
+          "tableId": "result_1"
+        }
       }
-    },
-    {
-      "type": "edge",
-      "edgeType": "order by",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_id"
-      },
-      "target": {
-        "tableId": "result_1"
-      }
-    }
-  ]
+    ]
+  }
 }
