+++ src/antlr/lineage.tests/PostgreSQL JSON operators.json
@@ -1,176 +1,180 @@
 {
   "sql": "select account_type::json->'b' as f1, '[1,2,3]'::json ->> 2 as f2, account_id#>account_type as f3, account_parent#>>account_id as f4 from account",
-  "data": [
-    {
-      "type": "table",
-      "id": "result_2",
-      "label": "account",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 138,
-        "endColumn": 145
-      },
-      "data": {
-        "id": "account"
-      },
-      "isSourceOnly": true,
-      "columns": [
-        {
-          "id": "account_type",
-          "label": "account_type",
-          "data": {
+  "data": {
+    "nodes": [
+      {
+        "type": "table",
+        "id": "result_2",
+        "label": "account",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 138,
+          "endColumn": 145
+        },
+        "data": {
+          "id": "account"
+        },
+        "isSourceOnly": true,
+        "columns": [
+          {
             "id": "account_type",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_id",
-          "label": "account_id",
-          "data": {
+            "label": "account_type",
+            "data": {
+              "id": "account_type",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_id",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_description",
-          "label": "account_description",
-          "data": {
+            "label": "account_id",
+            "data": {
+              "id": "account_id",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_description",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_parent",
-          "label": "account_parent",
-          "data": {
+            "label": "account_description",
+            "data": {
+              "id": "account_description",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_parent",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_rollup",
-          "label": "account_rollup",
-          "data": {
+            "label": "account_parent",
+            "data": {
+              "id": "account_parent",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_rollup",
-            "tableId": "account"
+            "label": "account_rollup",
+            "data": {
+              "id": "account_rollup",
+              "tableId": "account"
+            }
           }
+        ]
+      },
+      {
+        "type": "table",
+        "id": "result_1",
+        "label": "[final result]",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 0,
+          "endColumn": 145
+        },
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "f1",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 7,
+              "endColumn": 36
+            }
+          },
+          {
+            "id": "column_2",
+            "label": "f2",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 38,
+              "endColumn": 65
+            }
+          },
+          {
+            "id": "column_3",
+            "label": "f3",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 67,
+              "endColumn": 97
+            }
+          },
+          {
+            "id": "column_4",
+            "label": "f4",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 99,
+              "endColumn": 132
+            }
+          }
+        ],
+        "isTargetOnly": true
+      }
+    ],
+    "edges": [
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_type"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_1"
         }
-      ]
-    },
-    {
-      "type": "table",
-      "id": "result_1",
-      "label": "[final result]",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 0,
-        "endColumn": 145
       },
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "f1",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 7,
-            "endColumn": 36
-          }
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_id"
         },
-        {
-          "id": "column_2",
-          "label": "f2",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 38,
-            "endColumn": 65
-          }
-        },
-        {
-          "id": "column_3",
-          "label": "f3",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 67,
-            "endColumn": 97
-          }
-        },
-        {
-          "id": "column_4",
-          "label": "f4",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 99,
-            "endColumn": 132
-          }
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_3"
         }
-      ],
-      "isTargetOnly": true
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_type"
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_1"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_id"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_type"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_3"
+        }
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_3"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_type"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_parent"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_4"
+        }
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_3"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_id"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_4"
+        }
       }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_parent"
-      },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_4"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_id"
-      },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_4"
-      }
-    }
-  ]
+    ]
+  }
 }
