+++ src/antlr/lineage.tests/assumed columns on existing table.json
@@ -1,100 +1,104 @@
 {
   "sql": "SELECT account_id aid, account_details adet FROM account",
   "mergedLeaves": true,
-  "data": [
-    {
-      "type": "table",
-      "id": "result_2",
-      "label": "account",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 49,
-        "endColumn": 56
-      },
-      "data": {
-        "id": "account"
-      },
-      "isSourceOnly": true,
-      "columns": [
-        {
-          "id": "account_id",
-          "label": "account_id",
-          "data": {
+  "data": {
+    "nodes": [
+      {
+        "type": "table",
+        "id": "result_2",
+        "label": "account",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 49,
+          "endColumn": 56
+        },
+        "data": {
+          "id": "account"
+        },
+        "isSourceOnly": true,
+        "columns": [
+          {
             "id": "account_id",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "column_6",
-          "label": "account_details",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 23,
-            "endColumn": 38
+            "label": "account_id",
+            "data": {
+              "id": "account_id",
+              "tableId": "account"
+            }
           },
-          "isAssumed": true
-        }
-      ]
-    },
-    {
-      "type": "table",
-      "id": "result_1",
-      "label": "[final result]",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 0,
-        "endColumn": 56
-      },
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "aid",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 7,
-            "endColumn": 21
+          {
+            "id": "column_6",
+            "label": "account_details",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 23,
+              "endColumn": 38
+            },
+            "isAssumed": true
           }
+        ]
+      },
+      {
+        "type": "table",
+        "id": "result_1",
+        "label": "[final result]",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 0,
+          "endColumn": 56
         },
-        {
-          "id": "column_2",
-          "label": "adet",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 23,
-            "endColumn": 43
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "aid",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 7,
+              "endColumn": 21
+            }
+          },
+          {
+            "id": "column_2",
+            "label": "adet",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 23,
+              "endColumn": 43
+            }
           }
+        ],
+        "isTargetOnly": true
+      }
+    ],
+    "edges": [
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "account_id"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_1"
         }
-      ],
-      "isTargetOnly": true
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "account_id"
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_1"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "column_6"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_2"
+        }
       }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "column_6"
-      },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_2"
-      }
-    }
-  ]
+    ]
+  }
 }
