+++ src/antlr/lineage.tests/case sensitive insensitive identifiers.json
@@ -1,155 +1,159 @@
 {
   "sql": "select accOUNT_ID as \"acc_id\" from (select \"account_id\", ACCounT_TypE from account) as a",
-  "data": [
-    {
-      "type": "table",
-      "id": "result_3",
-      "label": "account",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 75,
-        "endColumn": 82
-      },
-      "data": {
-        "id": "account"
-      },
-      "isSourceOnly": true,
-      "columns": [
-        {
-          "id": "account_type",
-          "label": "account_type",
-          "data": {
+  "data": {
+    "nodes": [
+      {
+        "type": "table",
+        "id": "result_3",
+        "label": "account",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 75,
+          "endColumn": 82
+        },
+        "data": {
+          "id": "account"
+        },
+        "isSourceOnly": true,
+        "columns": [
+          {
             "id": "account_type",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_id",
-          "label": "account_id",
-          "data": {
+            "label": "account_type",
+            "data": {
+              "id": "account_type",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_id",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_description",
-          "label": "account_description",
-          "data": {
+            "label": "account_id",
+            "data": {
+              "id": "account_id",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_description",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_parent",
-          "label": "account_parent",
-          "data": {
+            "label": "account_description",
+            "data": {
+              "id": "account_description",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_parent",
-            "tableId": "account"
-          }
-        },
-        {
-          "id": "account_rollup",
-          "label": "account_rollup",
-          "data": {
+            "label": "account_parent",
+            "data": {
+              "id": "account_parent",
+              "tableId": "account"
+            }
+          },
+          {
             "id": "account_rollup",
-            "tableId": "account"
+            "label": "account_rollup",
+            "data": {
+              "id": "account_rollup",
+              "tableId": "account"
+            }
           }
-        }
-      ]
-    },
-    {
-      "type": "table",
-      "id": "result_2",
-      "label": "a",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 36,
-        "endColumn": 82
+        ]
       },
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "account_id",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 43,
-            "endColumn": 55
-          }
+      {
+        "type": "table",
+        "id": "result_2",
+        "label": "a",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 36,
+          "endColumn": 82
         },
-        {
-          "id": "column_2",
-          "label": "ACCounT_TypE",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 57,
-            "endColumn": 69
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "account_id",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 43,
+              "endColumn": 55
+            }
+          },
+          {
+            "id": "column_2",
+            "label": "ACCounT_TypE",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 57,
+              "endColumn": 69
+            }
           }
-        }
-      ]
-    },
-    {
-      "type": "table",
-      "id": "result_1",
-      "label": "[final result]",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 0,
-        "endColumn": 88
+        ]
       },
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "acc_id",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 7,
-            "endColumn": 29
+      {
+        "type": "table",
+        "id": "result_1",
+        "label": "[final result]",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 0,
+          "endColumn": 88
+        },
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "acc_id",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 7,
+              "endColumn": 29
+            }
           }
+        ],
+        "isTargetOnly": true
+      }
+    ],
+    "edges": [
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_3",
+          "columnId": "account_id"
+        },
+        "target": {
+          "tableId": "result_2",
+          "columnId": "column_1"
         }
-      ],
-      "isTargetOnly": true
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_3",
-        "columnId": "account_id"
       },
-      "target": {
-        "tableId": "result_2",
-        "columnId": "column_1"
-      }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_3",
-        "columnId": "account_type"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_3",
+          "columnId": "account_type"
+        },
+        "target": {
+          "tableId": "result_2",
+          "columnId": "column_2"
+        }
       },
-      "target": {
-        "tableId": "result_2",
-        "columnId": "column_2"
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "column_1"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_1"
+        }
       }
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "column_1"
-      },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_1"
-      }
-    }
-  ]
+    ]
+  }
 }
