+++ src/antlr/lineage.tests/derive fields for unknown tables.json
@@ -1,66 +1,70 @@
 {
   "sql": "SELECT a_column FROM a_table",
-  "data": [
-    {
-      "type": "table",
-      "id": "result_2",
-      "label": "a_table",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 21,
-        "endColumn": 28
-      },
-      "isSourceOnly": true,
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "a_column",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 7,
-            "endColumn": 15
-          },
-          "isAssumed": true
-        }
-      ]
-    },
-    {
-      "type": "table",
-      "id": "result_1",
-      "label": "[final result]",
-      "range": {
-        "startLine": 1,
-        "endLine": 1,
-        "startColumn": 0,
-        "endColumn": 28
-      },
-      "columns": [
-        {
-          "id": "column_1",
-          "label": "a_column",
-          "range": {
-            "startLine": 1,
-            "endLine": 1,
-            "startColumn": 7,
-            "endColumn": 15
+  "data": {
+    "nodes": [
+      {
+        "type": "table",
+        "id": "result_2",
+        "label": "a_table",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 21,
+          "endColumn": 28
+        },
+        "isSourceOnly": true,
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "a_column",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 7,
+              "endColumn": 15
+            },
+            "isAssumed": true
           }
-        }
-      ],
-      "isTargetOnly": true
-    },
-    {
-      "type": "edge",
-      "edgeType": "select",
-      "source": {
-        "tableId": "result_2",
-        "columnId": "column_1"
+        ]
       },
-      "target": {
-        "tableId": "result_1",
-        "columnId": "column_1"
+      {
+        "type": "table",
+        "id": "result_1",
+        "label": "[final result]",
+        "range": {
+          "startLine": 1,
+          "endLine": 1,
+          "startColumn": 0,
+          "endColumn": 28
+        },
+        "columns": [
+          {
+            "id": "column_1",
+            "label": "a_column",
+            "range": {
+              "startLine": 1,
+              "endLine": 1,
+              "startColumn": 7,
+              "endColumn": 15
+            }
+          }
+        ],
+        "isTargetOnly": true
       }
-    }
-  ]
+    ],
+    "edges": [
+      {
+        "type": "edge",
+        "edgeType": "select",
+        "source": {
+          "tableId": "result_2",
+          "columnId": "column_1"
+        },
+        "target": {
+          "tableId": "result_1",
+          "columnId": "column_1"
+        }
+      }
+    ]
+  }
 }
